pyproject.toml
[project]
name = "opti-bfm"
version = "0.1.0"
description = "Codebase for 'Optimistic Task Inference for Behavior Foundation Models'"
authors = [
    { name = "Thomas Rupf" },
    { name = "Marco Bagatella" },
]
maintainers = [
    { name = "Thomas Rupf", email = "thrupf@ethz.ch" }
]
license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: MIT License"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jax >= 0.6.1",
    "optax >= 0.2.4",
    "flax >= 0.10.6",
    "distrax >= 0.1.5",
    "hydra-core >= 1.3.2",
    "zarr >= 3.0.8",
    "wandb",
    "moviepy",
    "tqdm",
    "ogbench",
    "dm-control>=1.0.30",
    "numpy>=2.3.2",
]


[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.format]
quote-style = "single"
